Added grass
Also tried to fix the fps limit which I swear was working last time, not only working but working well. The actual frame rate was very inconsistent and was higher than what we wanted. By moving the delay after the update and draw calls it fixes it a bit but it's still fairly inconsistent.

+++ src/Game.h
@@ -38,7 +38,7 @@ public:
 
 	void LimitFPS(bool _limit) { mLimitingFPS = _limit; }
 	bool IsFPSLimited() { return mLimitingFPS; }
-	void SetFPSLimit(int _fps) { mFPSLimit = _fps; }
+	void SetFPSLimit(float _fps) { mFPSLimit = _fps; }
 	float GetFPSLimit() { return mFPSLimit; }
 
 	void StopGameRunning() { mGameRunning = false; }
@@ -78,7 +78,7 @@ private:
 	bool mSceneChanged = false;
 
 	bool mLimitingFPS = false;
-	int mFPSLimit = 60;
+	float mFPSLimit = 60;
 
 	// To detect changes for shader uniforms
 	glm::mat4 mLastProjection = glm::mat4(1.f);
